chore: move tumourSummaryType into common types
no-ticket
