+++ app/views/ReportView/components/GenomicSummary/index.tsx
@@ -27,7 +27,6 @@ import TumourSummaryEdit from '@/components/TumourSummaryEdit';
 import {
   PatientInformationType,
   GeneVariantType,
-  TumourSummaryType,
   MicrobialType,
 } from './types';
 import {
@@ -37,6 +36,7 @@ import MutationSignatureType from '../MutationSignatures/types';
 import { ImmuneType } from '../Immune/types';
 
 import './index.scss';
+import { TumourSummaryType } from '@/common';
 
 const variantCategory = (variant) => {
   // small mutations
